requirements.txt
absl-py~=1.1
fastapi~=0.86
httpx~=0.23
prometheus-client~=0.17
python-Levenshtein==0.21